Initial import of skeleton page for console hacking
Choose MIDI input and output ports from the drop-down selectors in the
header bar. Identifiers 'transmit' and 'receive' are exported in global
scope, accessible from the javascript console.

Call transmit() with an array of MIDI bytes to send that command to the
selected output. Set 'receive' to a function to be called with an array
of MIDI bytes for each command arriving on the selected input.