Replace the skeleton receive function with a handler system
Call receive() with a listener function and an optional timeout in
milliseconds to subscribe to inbound messages as arrays of MIDI bytes.
receive() is asynchronous, returning a promise which resolves when the
timeout expires or the listener returns true to signify completion.

Alternatively, add a handler manually to 'receivers' with
receivers.add() to subscribe to all message events, then remove it with
receivers.delete() when finished.