Support timestamps in skeleton transmit and receive
Pass a second optional argument to transmit() to schedule transmission
at the specified high-resolution time, or immediately if the time is in
the past. Listeners attached with receive() are called with both the
message data and the associated high-resolution time of receipt.

The high-resolution time stamps are floating point DOMHighResTimeStamp
values, measured in milliseconds relative to the start of navigation for
the document.