Initial import of Montage skeleton page
This page demonstrates automatic instrument detection, sending universal
identity requests on each available output, monitoring for corresponding
replies from a Yamaha Montage or MODX synthesizer.

The object 'montage' is exported in global scope, accessible from the
javascript console. This includes the relevant MIDI input port and
output port, instrument model name, firmware version and a series of
methods analogous to the functions exported by the MIDI skeleton page.

Scanning continues until a supported instrument is found, and a rescan
is automatically initiated when the instrument's ports are disconnected.
However, the instrument is not continuously polled when the detected
ports remain up.